requirements.txt
astroid==2.5.1
attrs==20.3.0
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
marshmallow==3.10.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
python-magic==0.4.22
requests==2.25.1
requests-toolbelt==0.9.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
